+++ tests/DFE.FindInformationAcademiesTrusts.CypressTests/cypress/e2e/regression/trusts/ofsted-page.cy.ts
@@ -2,8 +2,85 @@ import ofstedPage from "../../../pages/trusts/ofstedPage";
 import navigation from "../../../pages/navigation";
 import dataDownload from "../../../pages/trusts/dataDownload";
 import commonPage from "../../../pages/commonPage";
+import { TestDataStore } from "../../../support/test-data-store";
+
+const testTrustOfstedData = [
+    {
+        typeOfTrust: "single academy trust",
+        uid: 5527
+    },
+    {
+        typeOfTrust: "multi academy trust",
+        uid: 5712
+    }
+];
 
 describe("Testing the Ofsted page and its subpages ", () => {
+
+    testTrustOfstedData.forEach(({ typeOfTrust, uid }) => {
+        describe(`Testing the single headline grades page for a ${typeOfTrust}`, () => {
+            beforeEach(() => {
+                cy.login();
+                cy.visit(`/trusts/ofsted/single-headline-grades?uid=${uid}`);
+
+                cy.task('checkForFiles', 'cypress/downloads').then((files) => {
+                    if (files) {
+                        cy.task('clearDownloads', 'cypress/downloads');
+                    }
+                });
+            });
+
+            it("Checks the correct Ofsted single headline grades subpage header is present", () => {
+                ofstedPage
+                    .checkOfstedSHGPageHeaderPresent();
+            });
+
+            it("Checks the browser title is correct", () => {
+                commonPage
+                    .checkThatBrowserTitleForTrustPageMatches('Single headline grades - Ofsted - {trustName} - Find information about academies and trusts');
+            });
+
+            it("Checks the breadcrumb shows the correct page name", () => {
+                navigation
+                    .checkPageNameBreadcrumbPresent("Ofsted");
+            });
+
+            it("Checks the correct Ofsted single headline grades table headers are present", () => {
+                ofstedPage
+                    .checkOfstedSHGTableHeadersPresent();
+            });
+
+            it("Checks the Ofsted current ratings page sorting", () => {
+                ofstedPage
+                    .checkOfstedSHGSorting();
+            });
+
+            it("Checks that a trusts current and previous ratings correct judgement types are present", () => {
+                ofstedPage
+                    .checkSHGCurrentSHGJudgementsPresent()
+                    .checkSHGPreviousSHGJudgementsPresent()
+                    .checkSHGCurrentSHGBeforeOrAfterPresent()
+                    .checkSHGPreviousSHGBeforeOrAfterPresent();
+            });
+
+            it("Checks that the single headline grades dates are within the correct parameters", () => {
+                ofstedPage
+                    .checkSHGDateJoinedPresent()
+                    .checkSHGDateOfCurrentInspectionPresent()
+                    .checkSHGDateOfPreviousInspectionPresent();
+            });
+
+            it('should export academies data as an xlsx and verify it has downloaded and has content', () => {
+                ofstedPage
+                    .clickDownloadButton();
+                dataDownload
+                    .checkFileDownloaded()
+                    .checkFileHasContent()
+                    .deleteDownloadedFile();
+            });
+
+        });
+    });
 
     describe("Testing the Ofsted current ratings page ", () => {
         beforeEach(() => {
@@ -207,77 +284,15 @@ describe("Testing the Ofsted page and its subpages ", () => {
         });
     });
 
-    describe("Testing the Ofsted important dates page ", () => {
-        beforeEach(() => {
-            cy.login();
-            cy.visit('/trusts/ofsted/important-dates?uid=5143');
-
-            // Clear the downloads folder before running each test
-            cy.task('checkForFiles', 'cypress/downloads').then((files) => {
-                if (files) {
-                    cy.task('clearDownloads', 'cypress/downloads');
-                }
-            });
-        });
-
-        it("Checks the correct Ofsted important dates sub page header is present", () => {
-            ofstedPage
-                .checkOfstedImportantDatesPageHeaderPresent();
-        });
-
-        it("Checks the browser title is correct", () => {
-            commonPage
-                .checkThatBrowserTitleForTrustPageMatches('Important dates - Ofsted - {trustName} - Find information about academies and trusts');
-        });
-
-        it("Checks the breadcrumb shows the correct page name", () => {
-            navigation
-                .checkPageNameBreadcrumbPresent("Ofsted");
-        });
-
-        it("Checks the correct Ofsted important dates table headers are present", () => {
-            ofstedPage
-                .checkOfstedImportantDatesTableHeadersPresent();
-        });
-
-        it("Checks that a trusts important dates fields are present ", () => {
-            ofstedPage
-                .checkDateJoinedPresent()
-                .checkDateOfCurrentInspectionPresent()
-                .checkDateOfPreviousInspectionPresent();
-        });
-
-        it("Checks that a trusts important dates sorting is working", () => {
-            ofstedPage
-                .checkOfstedImportantDatesSorting();
-        });
-
-        it("Checks that a different trusts important dates fields are present", () => {
-            cy.visit('/trusts/ofsted/important-dates?uid=5712');
-            ofstedPage
-                .checkDateJoinedPresent()
-                .checkDateOfCurrentInspectionPresent()
-                .checkDateOfPreviousInspectionPresent();
-        });
-
-        it('should export academies data as an xlsx and verify it has downloaded and has content', () => {
-            ofstedPage
-                .clickDownloadButton();
-            dataDownload
-                .checkFileDownloaded()
-                .checkFileHasContent()
-                .deleteDownloadedFile();
-        });
-    });
-
     describe("Testing a trust that has no ofsted data within it to ensure the issue of a 500 page appearing does not happen", () => {
         beforeEach(() => {
             cy.login();
-            commonPage.interceptAndVerifyNo500Errors();
+            commonPage
+                .interceptAndVerifyNo500Errors();
         });
 
-        ['/trusts/ofsted/current-ratings?uid=17728', '/trusts/ofsted/previous-ratings?uid=17728', '/trusts/ofsted/important-dates?uid=17728', '/trusts/ofsted/safeguarding-and-concerns?uid=17728'].forEach((url) => {
-            it(`Should have no 500 error on ${url}`, () => {
+        TestDataStore.GetTrustSubpagesFor(17728, "Ofsted").forEach(({ subpageName, url }) => {
+            it(`Should have no 500 error on ${subpageName}`, () => {
                 cy.visit(url);
             });
         });