+++ tests/DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.UnitTests/Repositories/AcademyRepositoryOfstedTests.cs
@@ -476,7 +476,8 @@ public class AcademyRepositoryOfstedTests
         // The first three urns are set up in non-further establishments with a non-further only property
         var nonFurtherUrns = allUrns.Take(3).ToArray();
         _mockAcademiesDbContext.AddEstablishmentsFiat(
-            nonFurtherUrns.Select(urn => new MisMstrEstablishmentFiat { Urn = urn, EarlyYearsProvisionWhereApplicable = 1 }));
+            nonFurtherUrns.Select(urn => new MisMstrEstablishmentFiat
+                { Urn = urn, EarlyYearsProvisionWhereApplicable = 1 }));
 
         // All urns are set up in further (note that this wouldn't occur in the actual db)
         _mockAcademiesDbContext.AddFurtherEducationEstablishmentsFiat(
@@ -567,7 +568,7 @@ public class AcademyRepositoryOfstedTests
         var result = AcademyRepository.ConvertOverallEffectivenessToOfstedRatingScore(rating);
 
         // Assert
-        result.Should().Be(OfstedRatingScore.NoJudgement);
+        result.Should().Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
     }
 
     [Theory]
@@ -576,7 +577,7 @@ public class AcademyRepositoryOfstedTests
     [InlineData("3", OfstedRatingScore.RequiresImprovement)]
     [InlineData("4", OfstedRatingScore.Inadequate)]
     [InlineData("8", OfstedRatingScore.DoesNotApply)]
-    [InlineData("9", OfstedRatingScore.NoJudgement)]
+    [InlineData("9", OfstedRatingScore.SingleHeadlineGradeNotAvailable)]
     [InlineData("0", OfstedRatingScore.InsufficientEvidence)]
     [InlineData("-1", OfstedRatingScore.NotInspected)]
     public void
@@ -625,11 +626,12 @@ public class AcademyRepositoryOfstedTests
 
     [Theory]
     [InlineData("1", "2", OfstedRatingScore.Outstanding, OfstedRatingScore.Good)]
-    [InlineData("Not judged", "Not judged", OfstedRatingScore.NoJudgement, OfstedRatingScore.NoJudgement)]
+    [InlineData("Not judged", "Not judged", OfstedRatingScore.SingleHeadlineGradeNotAvailable,
+        OfstedRatingScore.SingleHeadlineGradeNotAvailable)]
     [InlineData(null, null, OfstedRatingScore.NotInspected, OfstedRatingScore.NotInspected)]
     [InlineData("abc", "def", OfstedRatingScore.Unknown, OfstedRatingScore.Unknown)]
-    [InlineData("2", "Not judged", OfstedRatingScore.Good, OfstedRatingScore.NoJudgement)]
-    [InlineData("Not judged", "1", OfstedRatingScore.NoJudgement, OfstedRatingScore.Outstanding)]
+    [InlineData("2", "Not judged", OfstedRatingScore.Good, OfstedRatingScore.SingleHeadlineGradeNotAvailable)]
+    [InlineData("Not judged", "1", OfstedRatingScore.SingleHeadlineGradeNotAvailable, OfstedRatingScore.Outstanding)]
     public async Task
         GetAcademiesInTrustOfstedAsync_should_correctly_convert_OverallEffectiveness_and_PreviousFullInspectionOverallEffectiveness(
             string? overallEffectiveness,
@@ -658,5 +660,270 @@ public class AcademyRepositoryOfstedTests
         academyOfsted.Urn.Should().Be(giasGroupLink.Urn);
         academyOfsted.CurrentOfstedRating.OverallEffectiveness.Should().Be(expectedCurrentScore);
         academyOfsted.PreviousOfstedRating.OverallEffectiveness.Should().Be(expectedPreviousScore);
+    }
+
+    [Fact]
+    public async Task
+        GetAcademiesInTrustOfstedAsync_should_remove_single_headline_grades_issued_after_2nd_sept_2024()
+    {
+        // Arrange
+        var giasGroupLinks = AddGiasGroupLinksToMockDb(10);
+
+        // - Add non-further eds
+        _mockAcademiesDbContext.AddEstablishmentsFiat(
+            new MisMstrEstablishmentFiat // OverallEffectiveness set after policy change date
+            {
+                Urn = int.Parse(giasGroupLinks[0].Urn!),
+                OverallEffectiveness = "1",
+                PreviousFullInspectionOverallEffectiveness = "3",
+                InspectionStartDate = "01/01/2025",
+                PreviousInspectionStartDate = "01/01/2021"
+            },
+            new MisMstrEstablishmentFiat // OverallEffectiveness set on policy change date
+            {
+                Urn = int.Parse(giasGroupLinks[1].Urn!),
+                OverallEffectiveness = "2",
+                InspectionStartDate = "02/09/2024"
+            },
+            new MisMstrEstablishmentFiat // PreviousFullInspectionOverallEffectiveness set after policy change date
+            {
+                Urn = int.Parse(giasGroupLinks[2].Urn!),
+                OverallEffectiveness = "Not judged",
+                PreviousFullInspectionOverallEffectiveness = "3",
+                InspectionStartDate = "01/01/2025",
+                PreviousInspectionStartDate = "12/12/2024"
+            },
+            new MisMstrEstablishmentFiat // PreviousFullInspectionOverallEffectiveness set on policy change date
+            {
+                Urn = int.Parse(giasGroupLinks[3].Urn!),
+                OverallEffectiveness = "Not judged",
+                PreviousFullInspectionOverallEffectiveness = "4",
+                InspectionStartDate = "01/01/2025",
+                PreviousInspectionStartDate = "02/09/2024"
+            },
+            new MisMstrEstablishmentFiat // OverallEffectiveness before policy change date
+            {
+                Urn = int.Parse(giasGroupLinks[4].Urn!),
+                OverallEffectiveness = "1",
+                PreviousFullInspectionOverallEffectiveness = "2",
+                InspectionStartDate = "01/09/2024",
+                PreviousInspectionStartDate = "01/01/2021"
+            }
+        );
+
+        // - Add further eds
+        _mockAcademiesDbContext.AddFurtherEducationEstablishmentsFiat(
+            new MisMstrFurtherEducationEstablishmentFiat // OverallEffectiveness set after policy change date
+            {
+                ProviderUrn = int.Parse(giasGroupLinks[5].Urn!),
+                OverallEffectiveness = "1",
+                PreviousOverallEffectiveness = "3",
+                LastDayOfInspection = "01/01/2025",
+                PreviousLastDayOfInspection = "01/01/2021"
+            },
+            new MisMstrFurtherEducationEstablishmentFiat // OverallEffectiveness set on policy change date
+            {
+                ProviderUrn = int.Parse(giasGroupLinks[6].Urn!),
+                OverallEffectiveness = "2",
+                LastDayOfInspection = "02/09/2024"
+            },
+            new
+                MisMstrFurtherEducationEstablishmentFiat // PreviousFullInspectionOverallEffectiveness set after policy change date
+                {
+                    ProviderUrn = int.Parse(giasGroupLinks[7].Urn!),
+                    OverallEffectiveness = "Not judged",
+                    PreviousOverallEffectiveness = "3",
+                    LastDayOfInspection = "01/01/2025",
+                    PreviousLastDayOfInspection = "12/12/2024"
+                },
+            new
+                MisMstrFurtherEducationEstablishmentFiat // PreviousFullInspectionOverallEffectiveness set on policy change date
+                {
+                    ProviderUrn = int.Parse(giasGroupLinks[8].Urn!),
+                    OverallEffectiveness = "Not judged",
+                    PreviousOverallEffectiveness = "4",
+                    LastDayOfInspection = "01/01/2025",
+                    PreviousLastDayOfInspection = "02/09/2024"
+                },
+            new MisMstrFurtherEducationEstablishmentFiat // OverallEffectiveness before policy change date
+            {
+                ProviderUrn = int.Parse(giasGroupLinks[9].Urn!),
+                OverallEffectiveness = "1",
+                PreviousOverallEffectiveness = "2",
+                LastDayOfInspection = "01/09/2024",
+                PreviousLastDayOfInspection = "01/01/2021"
+            }
+        );
+
+        // Act
+        var result = await _sut.GetAcademiesInTrustOfstedAsync(GroupUid);
+
+        // Assert
+        result.Should().SatisfyRespectively(
+            // Non-further eds
+            rating => // OverallEffectiveness set after policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .NotBe(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+            rating => // OverallEffectiveness set on policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .NotBe(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+            rating => // PreviousFullInspectionOverallEffectiveness set after policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+            rating => // PreviousFullInspectionOverallEffectiveness set on policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+            rating => // OverallEffectiveness before policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .NotBe(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .NotBe(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+
+            // Further eds
+            rating => // OverallEffectiveness set after policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .NotBe(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+            rating => // OverallEffectiveness set on policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .NotBe(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+            rating => // PreviousFullInspectionOverallEffectiveness set after policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+            rating => // PreviousFullInspectionOverallEffectiveness set on policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .Be(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            },
+            rating => // OverallEffectiveness before policy change date
+            {
+                rating.CurrentOfstedRating.OverallEffectiveness.Should()
+                    .NotBe(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+                rating.PreviousOfstedRating.OverallEffectiveness.Should()
+                    .NotBe(OfstedRatingScore.SingleHeadlineGradeNotAvailable);
+            }
+        );
+
+        //Assert error logging occurred
+        _mockLogger.VerifyLogErrors(
+            $"URN {giasGroupLinks[0].Urn!} has a current Ofsted single headline grade of Outstanding issued",
+            $"URN {giasGroupLinks[1].Urn!} has a current Ofsted single headline grade of Good issued",
+            $"URN {giasGroupLinks[2].Urn!} has a previous Ofsted single headline grade of RequiresImprovement issued",
+            $"URN {giasGroupLinks[3].Urn!} has a previous Ofsted single headline grade of Inadequate issued",
+            // giasGroupLinks[4] - OverallEffectiveness before policy change date so no error log expected
+            $"URN {giasGroupLinks[5].Urn!} has a current Ofsted single headline grade of Outstanding issued",
+            $"URN {giasGroupLinks[6].Urn!} has a current Ofsted single headline grade of Good issued",
+            $"URN {giasGroupLinks[7].Urn!} has a previous Ofsted single headline grade of RequiresImprovement issued",
+            $"URN {giasGroupLinks[8].Urn!} has a previous Ofsted single headline grade of Inadequate issued"
+            // giasGroupLinks[9] - OverallEffectiveness before policy change date so no error log expected
+        );
+        _mockLogger.VerifyNoOtherCalls();
+    }
+
+    [Fact]
+    public async Task GetAcademiesInTrustOfstedAsync_should_not_log_error_for_valid_overall_effectiveness()
+    {
+        // Arrange
+        var giasGroupLinks = AddGiasGroupLinksToMockDb(8);
+
+        // - Add non-further eds
+        _mockAcademiesDbContext.AddEstablishmentsFiat(
+            new MisMstrEstablishmentFiat // OverallEffectiveness before policy change date
+            {
+                Urn = int.Parse(giasGroupLinks[0].Urn!),
+                OverallEffectiveness = "1",
+                PreviousFullInspectionOverallEffectiveness = "2",
+                InspectionStartDate = "01/09/2024",
+                PreviousInspectionStartDate = "01/01/2021"
+            },
+            new MisMstrEstablishmentFiat // Not judged on policy change date
+            {
+                Urn = int.Parse(giasGroupLinks[1].Urn!),
+                OverallEffectiveness = "Not judged",
+                PreviousFullInspectionOverallEffectiveness = "2",
+                InspectionStartDate = "02/09/2024",
+                PreviousInspectionStartDate = "01/01/2021"
+            },
+            new MisMstrEstablishmentFiat // Not judged after policy change date
+            {
+                Urn = int.Parse(giasGroupLinks[2].Urn!),
+                OverallEffectiveness = "Not judged",
+                PreviousFullInspectionOverallEffectiveness = "Not judged",
+                InspectionStartDate = "20/05/2025",
+                PreviousInspectionStartDate = "02/09/2024"
+            },
+            new MisMstrEstablishmentFiat // Not inspected
+            {
+                Urn = int.Parse(giasGroupLinks[3].Urn!)
+            }
+        );
+
+        // - Add further eds
+        _mockAcademiesDbContext.AddFurtherEducationEstablishmentsFiat(
+            new MisMstrFurtherEducationEstablishmentFiat // OverallEffectiveness before policy change date
+            {
+                ProviderUrn = int.Parse(giasGroupLinks[4].Urn!),
+                OverallEffectiveness = "1",
+                PreviousOverallEffectiveness = "2",
+                LastDayOfInspection = "01/09/2024",
+                PreviousLastDayOfInspection = "01/01/2021"
+            },
+            new MisMstrFurtherEducationEstablishmentFiat // Not judged on policy change date
+            {
+                ProviderUrn = int.Parse(giasGroupLinks[5].Urn!),
+                OverallEffectiveness = "Not judged",
+                PreviousOverallEffectiveness = "2",
+                LastDayOfInspection = "02/09/2024",
+                PreviousLastDayOfInspection = "01/01/2021"
+            },
+            new MisMstrFurtherEducationEstablishmentFiat // Not judged after policy change date
+            {
+                ProviderUrn = int.Parse(giasGroupLinks[6].Urn!),
+                OverallEffectiveness = "Not judged",
+                PreviousOverallEffectiveness = "Not judged",
+                LastDayOfInspection = "20/05/2025",
+                PreviousLastDayOfInspection = "02/09/2024"
+            },
+            new MisMstrFurtherEducationEstablishmentFiat // Not inspected
+            {
+                ProviderUrn = int.Parse(giasGroupLinks[7].Urn!)
+            }
+        );
+
+        // Act
+        _ = await _sut.GetAcademiesInTrustOfstedAsync(GroupUid);
+
+        _mockLogger.VerifyNoOtherCalls();
     }
 }