+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/Trusts/Ofsted/BaseOfstedAreaModelTests.cs
@@ -25,6 +25,12 @@ public abstract class BaseOfstedAreaModelTests<T> : BaseTrustPageTests<T>, ITest
         MockDataSourceService.Verify(e => e.GetAsync(Source.Mis), Times.Once);
 
         Sut.DataSourcesPerPage.Should().BeEquivalentTo([
+            new DataSourcePageListEntry(ViewConstants.OfstedSingleHeadlineGradesPageName, [
+                    new DataSourceListEntry(GiasDataSource, "Date joined trust"),
+                    new DataSourceListEntry(MisDataSource, "All single headline grades"),
+                    new DataSourceListEntry(MisDataSource, "All inspection dates")
+                ]
+            ),
             new DataSourcePageListEntry(ViewConstants.OfstedCurrentRatingsPageName, [
                     new DataSourceListEntry(MisDataSource, "Current Ofsted rating"),
                     new DataSourceListEntry(MisDataSource, "Date of current inspection")
@@ -32,12 +38,6 @@ public abstract class BaseOfstedAreaModelTests<T> : BaseTrustPageTests<T>, ITest
             ),
             new DataSourcePageListEntry(ViewConstants.OfstedPreviousRatingsPageName, [
                     new DataSourceListEntry(MisDataSource, "Previous Ofsted rating"),
-                    new DataSourceListEntry(MisDataSource, "Date of previous inspection")
-                ]
-            ),
-            new DataSourcePageListEntry(ViewConstants.OfstedImportantDatesPageName, [
-                    new DataSourceListEntry(GiasDataSource, "Date joined trust"),
-                    new DataSourceListEntry(MisDataSource, "Date of current inspection"),
                     new DataSourceListEntry(MisDataSource, "Date of previous inspection")
                 ]
             ),
@@ -159,6 +159,12 @@ public abstract class BaseOfstedAreaModelTests<T> : BaseTrustPageTests<T>, ITest
             .SatisfyRespectively(
                 l =>
                 {
+                    l.LinkText.Should().Be("Single headline grades");
+                    l.SubPageLink.Should().Be("./SingleHeadlineGrades");
+                    l.ServiceName.Should().Be("Ofsted");
+                },
+                l =>
+                {
                     l.LinkText.Should().Be("Current ratings");
                     l.SubPageLink.Should().Be("./CurrentRatings");
                     l.ServiceName.Should().Be("Ofsted");
@@ -167,12 +173,6 @@ public abstract class BaseOfstedAreaModelTests<T> : BaseTrustPageTests<T>, ITest
                 {
                     l.LinkText.Should().Be("Previous ratings");
                     l.SubPageLink.Should().Be("./PreviousRatings");
-                    l.ServiceName.Should().Be("Ofsted");
-                },
-                l =>
-                {
-                    l.LinkText.Should().Be("Important dates");
-                    l.SubPageLink.Should().Be("./ImportantDates");
                     l.ServiceName.Should().Be("Ofsted");
                 },
                 l =>